Remove home state and abolish fixture-attribute use in states
The "home state" has no reason to exist.  The home values are already
stored perfectly well in the attribute lists of the fixtures.  Any time
we need to look up a home value, we already have the fixture itself
available.

This also gets rid of any use of <fixture-attribute> in states.  This
was confusing me.  Better to just pass symbols around and only get the
real attribute objects when needed (which isn't very often).