Fixture display: Avoid hard transition of intensity text colour
The snap transition is bad, because it gives the impression that some
kind of snap change has taken place in the lighting state.

Since the background has to get lighter and the text darker, they have
to cross over (making the text invisible) at some value.  This is an
attempt to make the crossover time as short as possible without any
snap transition.